+++ param_global/hooks.py
@@ -36,6 +36,7 @@ app_include_js = [
 	"pg_export_defaults.bundle.js",
 	"pg_link_search.bundle.js",
 	"pg_nombres.bundle.js",
+	"pg_verrou.bundle.js",
 ]
 
 # include js, css files in header of web template